% Generated by roxygen2 (4.0.2): do not edit by hand
\name{all.part.reg}
\alias{all.part.reg}
\title{Defining the candidates to the optimum partition for each of segmentation variables
(regression model)}
\usage{
\method{all}{part.reg}(x, y, method, ...)
}
\arguments{
\item{x}{matrix or data.frame with the data.}

\item{y}{matrix or data.frame of the segmentation variables.}

\item{method}{string indicating the method: LM or LAD}

\item{\dots}{Further arguments passed on to \code{\link{all.part.reg}}.}
}
\value{
list containing information of the candidates to the optimum partition for each of segmentation variables
}
\description{
Defining the candidates to the optimum partition for each of segmentation variables
(regression model)
}
\details{
Internal function. \code{all.part.reg} is called by \code{partopt.reg}.
}
\keyword{internal}

