% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_fun_blockdiag.R
\name{blockdiag}
\alias{blockdiag}
\title{Building the Bart matrix}
\usage{
blockdiag(x, ...)
}
\arguments{
\item{x}{list of matrix used to build the Bart matrix}

\item{\dots}{Further arguments passed on to \code{\link{blockdiag}}.}
}
\value{
Bart matrix
}
\description{
Building the Bart matrix
}
\details{
Internal function. \code{blockdiag} is called by \code{build.block}.
}
\keyword{internal}
