% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_fun_build.block.R
\name{build.block}
\alias{build.block}
\title{Defining the linear relations between the latent variables.}
\usage{
build.block(inner, latent, ...)
}
\arguments{
\item{inner}{matrix containing information of the latent causal relationships.}

\item{latent}{matrix containing the estimated latent variables.}

\item{\dots}{Further arguments passed on to \code{\link{build.block}}.}
}
\value{
list of matrices containing for each dependend latent variable the dependent LV and the explicative latent variables.
}
\description{
Defining the linear relations between the latent variables.
}
\details{
Internal function. \code{build.block} is called by \code{splitopt.pls}.
}
\keyword{internal}
