% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal_tree.R
\name{candidates.tree}
\alias{candidates.tree}
\title{Posibble partions for each node of the tree}
\usage{
candidates.tree(moxtree)
}
\arguments{
\item{moxtree}{class containing the moxtree elements}
}
\value{
the Posibble partions for each node of the tree
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
