% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathmox_internal.R
\name{check_arg_mox}
\alias{check_arg_mox}
\title{Checks arguments}
\usage{
check_arg_mox(x, y, signif, size, deep)
}
\arguments{
\item{x}{matrix or dataframe containing the data}

\item{y}{vector or dataframe containing the categorical variables}

\item{signif}{minimum threshold of f-test p-value}

\item{size}{minimum threshold size node}

\item{deep}{minimum threshold of deep tree}
}
\value{
a list checked arguments
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
