% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{check_boot}
\alias{check_boot}
\title{Check bootstrap options}
\usage{
check_boot(boot.val, br)
}
\arguments{
\item{boot.val}{logical indicating whether to perform bootstrapping}

\item{br}{number of bootstrap resamples}
}
\value{
validated bootstrap options
}
\description{
Check bootstrap options
}
\details{
Internal function. \code{check_boot} is called by \code{check_args}.
}
\keyword{internal}
