% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathmox_internal.R
\name{check_const}
\alias{check_const}
\title{Check consistence}
\usage{
check_const(.model, .data)
}
\arguments{
\item{.model}{a model in lavaan model syntax}

\item{.data}{a data.frame or a matrix of indicators}
}
\value{
a logical value
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
