% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{check_data}
\alias{check_data}
\title{Check Data}
\usage{
check_data(Data)
}
\arguments{
\item{Data}{numeric matrix or data frame containing the manifest variables.}
}
\value{
validated Data
}
\description{
Check Data
}
\details{
Internal function. \code{check_data} is called by \code{check_args}.
}
\keyword{internal}
