% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{check_maxiter}
\alias{check_maxiter}
\title{Check maximum number of iterations}
\usage{
check_maxiter(maxiter)
}
\arguments{
\item{maxiter}{integer indicating the maximum number of iterations
(iterative stage of the PLS algorithm)}
}
\value{
validated maxiter
}
\description{
Check maximum number of iterations
}
\details{
Internal function. \code{check_maxiter} is called by \code{check_specs}.
}
\keyword{internal}
