% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{check_model}
\alias{check_model}
\title{Check congruence between inner and outer models}
\usage{
check_model(path_matrix, blocks)
}
\arguments{
\item{path_matrix}{matrix specifying the path connections}

\item{blocks}{list defining the blocks of manifest variables}
}
\value{
an error if there is something wrong
}
\description{
Check congruence between inner and outer models
}
\details{
Internal function. \code{check_model} is called by \code{check_args}.
}
\keyword{internal}
