% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{check_path}
\alias{check_path}
\title{Check path matrix}
\usage{
check_path(path_matrix)
}
\arguments{
\item{path_matrix}{square (lower triangular) boolean matrix for inner model}
}
\value{
validated path matrix
}
\description{
Check path matrix
}
\details{
Internal function. \code{check_path} is called by \code{check_args}.
}
\keyword{internal}
