% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{check_scaling}
\alias{check_scaling}
\title{Check types of measurement scales and metric}
\usage{
check_scaling(scaling, scaled, blocks)
}
\arguments{
\item{scaling}{list with type of scaling for each manifest variable}

\item{scaled}{should manifest variables be standardized}

\item{blocks}{list defining the blocks of manifest variables}
}
\value{
list with validated scaling and scaled
}
\description{
Check types of measurement scales and metric
}
\details{
Internal function. \code{check_scaling} is called by \code{check_specs}.
}
\keyword{internal}
