% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{check_scheme}
\alias{check_scheme}
\title{Check scheme}
\usage{
check_scheme(scheme)
}
\arguments{
\item{scheme}{character indicating the inner weighting scheme}
}
\value{
validated scheme
}
\description{
Check scheme
}
\details{
Internal function. \code{check_scheme} is called by \code{check_specs}.
}
\keyword{internal}
