% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{check_tol}
\alias{check_tol}
\title{Check tolerance threshold}
\usage{
check_tol(tol)
}
\arguments{
\item{tol}{decimal value indicating the tolerance threshold for the 
convergence of the PLS algorithm (iterative stage)}
}
\value{
validated tol
}
\description{
Check tolerance threshold
}
\details{
Internal function. \code{check_tol} is called by \code{check_specs}.
}
\keyword{internal}
