% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathmox_internal.R
\name{comb}
\alias{comb}
\title{Combinations of a vector element}
\usage{
comb(n, r, v = 1:n, set = TRUE, repeats.allowed = FALSE)
}
\arguments{
\item{n}{size of the source vector}

\item{r}{size of the target vectors}

\item{v}{source vector. Defaults to 1:n}

\item{set}{logical flag indicating whether duplicates should be removed from the
source vector v}

\item{repeats.allowed}{logical flag indicating whether the constructed vectors may include
duplicated values}
}
\value{
a matrix of combinations
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
