% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_turner.R
\name{df_to_blocks}
\alias{df_to_blocks}
\title{Split a data frame into blocks}
\usage{
df_to_blocks(DataFrame, blocks, byrow = TRUE)
}
\arguments{
\item{DataFrame}{a data frame to split}

\item{blocks}{either a list or a vector indicating the blocks. If 
\code{blocks} is a list of vectors, then the length of each vector defines 
the size of the blocks. If \code{blocks} is a vector, then each element 
represents the size of the blocks.}

\item{byrow}{logical. If \code{TRUE} (the default) the data frame 
is split by rows, otherwise the data frame is split by columns}
}
\value{
A list of data frames
}
\description{
Split a data frame into blocks
}
\details{
Internal function.
}
\keyword{internal}
