% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathmox_internal.R
\name{.path}
\alias{.path}
\title{Path coefficient extraction}
\usage{
.path(x, l)
}
\arguments{
\item{x}{matrix containing path coefficients}
}
\value{
a vector of path coefficients
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
