% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{dummy.G}
\alias{dummy.G}
\title{Dummy by Giorgio}
\usage{
dummy.G(Y, X)
}
\arguments{
\item{Y}{a matrix}

\item{X}{a matrix}
}
\value{
list with Xdummy, Quant, eta2
}
\description{
Dummy by Giorgio
}
\details{
Internal function. \code{dummy.G} is called by \code{get_PLSRdoubleQ}.
}
\keyword{internal}
