% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_turner.R
\name{dummy_to_list}
\alias{dummy_to_list}
\title{Create an indexed list from a dummy matrix}
\usage{
dummy_to_list(Dummy)
}
\arguments{
\item{Dummy}{matrix (dummy by columns)}
}
\value{
A list of indexed vectors
}
\description{
Create an indexed list from a dummy matrix
}
\details{
Internal function.
}
\keyword{internal}
