% Generated by roxygen2 (4.0.2): do not edit by hand
\name{F.data.reg}
\alias{F.data.reg}
\title{Defining the matrices needed for the comparison test (regression model)}
\usage{
F.data.reg(x, modtwo, ...)
}
\arguments{
\item{x}{matrix or data.frame with the data.}

\item{modtwo}{vector indicating the binary partition.}

\item{\dots}{Further arguments passed on to \code{\link{F.data.reg}}.}
}
\value{
list containing matrices needed for the comparison test
}
\description{
Defining the matrices needed for the comparison test (regression model)
}
\details{
Internal function. \code{F.data.reg} is called by \code{test.partition.reg}.
}
\keyword{internal}

