% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_fun_f.min.R
\name{f.min}
\alias{f.min}
\title{Defining vector minimum position}
\usage{
f.min(x, ...)
}
\arguments{
\item{x}{vector of value.}

\item{\dots}{Further arguments passed on to \code{\link{f.min}}.}
}
\value{
list containing minimun value, the position of the minimun and the values different from NA of a vactor.
}
\description{
Defining vector minimum position
}
\details{
Internal function. \code{f.min} is called by \code{splitopt.pls}.
}
\keyword{internal}
