% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_turner.R
\name{factor_to_dummy}
\alias{factor_to_dummy}
\title{Create a dummy matrix from the elements in a factor}
\usage{
factor_to_dummy(afactor)
}
\arguments{
\item{afactor}{a factor (preferably of vectors)}
}
\value{
A matrix of dummy variables
}
\description{
Create a dummy matrix from the elements in a factor
}
\details{
Internal function.
}
\keyword{internal}
