% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Fb.test.pls}
\alias{Fb.test.pls}
\title{Block test}
\usage{
Fb.test.pls(Y1, X1, info.block, method, ...)
}
\arguments{
\item{Y1}{vector of the concatenate indipendent latent variables of alternative hypothesis block test.}

\item{X1}{matrix of the concatenate predictor latent variables of alternative hypothesis block test.}

\item{info.block}{list contaning information about the endogenous equations of the pls model.}

\item{method}{string indicating the method: LM or LAD.}

\item{\dots}{Further arguments passed on to \code{\link{Fb.test.pls}}.}
}
\value{
list containing the statistic and the p-value of block test
}
\description{
Block test
}
\details{
Internal function. \code{Fb.test.pls} is called by \code{test.partition.pls}.
}
\keyword{internal}

