% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23_fun_fblock.tree.pls.R
\name{fblock.tree.pls}
\alias{fblock.tree.pls}
\title{Listing the the f-block test results for each obtained partition}
\usage{
fblock.tree.pls(tree, ...)
}
\arguments{
\item{tree}{class containing the tree element.}

\item{\dots}{Further arguments passed on to \code{\link{fblock.tree.pls}}.}
}
\value{
the f-block test results for each obtained partition
}
\description{
Listing the the f-block test results for each obtained partition
}
\details{
Internal function. \code{fblock.tree.pls} is called by \code{pls.pathmox}.
}
\keyword{internal}
