% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Fc.test.pls}
\alias{Fc.test.pls}
\title{Coefficient test}
\usage{
Fc.test.pls(Y1, X1, path.name, info.block, method, ...)
}
\arguments{
\item{Y1}{vector of the concatenate indipendent latent variables of alternative hypothesis coefficient test.}

\item{X1}{matrix of the concatenate predictor latent variables of alternative hypothesis coefficient test.}

\item{path.name}{vector of path coefficients labels.}

\item{info.block}{list contaning information about the endogenous equations of the pls model.}

\item{method}{string indicating the method: LM or LAD.}

\item{\dots}{Further arguments passed on to \code{\link{Fc.test.pls}}.}
}
\value{
list containing the statistic and the p-value of coefficient test
}
\description{
Coefficient test
}
\details{
Internal function. \code{Fc.test.pls} is called by \code{test.partition.pls}.
}
\keyword{internal}

