% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal_tree.R
\name{fcoef.tree}
\alias{fcoef.tree}
\title{F-coefficients test results for each tree partition}
\usage{
fcoef.tree(moxtree)
}
\arguments{
\item{moxtree}{class containing the moxtree elements}
}
\value{
the F-coefficients test results for each tree partition
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
