% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fcoef.tree.reg}
\alias{fcoef.tree.reg}
\title{Listing the the f-coefficients test results for regression model for each obtained partition}
\usage{
fcoef.tree.reg(tree, ...)
}
\arguments{
\item{tree}{class containing the tree element.}

\item{\dots}{Further arguments passed on to \code{\link{fcoef.tree.reg}}.}
}
\value{
the f-coefficients test results for each obtained partition
}
\description{
Listing the the f-coefficients test results for regression model for each obtained partition
}
\details{
Internal function. \code{fcoef.tree.reg} is called by \code{reg.pathmox}.
}
\keyword{internal}

