% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_fun_Fg.test.pls.R
\name{Fg.test.pls}
\alias{Fg.test.pls}
\title{F-Global test.}
\usage{
Fg.test.pls(Y0, X0, Y1, X1, method, ...)
}
\arguments{
\item{Y0}{vector of the concatenate indipendent latent variables of null hypothesis global test.}

\item{X0}{matrix of the concatenate predictor latent variables of null hypothesis global test.}

\item{Y1}{vector of the concatenate indipendent latent variables of alternative hypothesis global test.}

\item{X1}{matrix of the concatenate predictor latent variables of alternative hypothesis global test.}

\item{method}{string indicating the method: LM or LAD}

\item{\dots}{Further arguments passed on to \code{\link{Fg.test.pls}}.}
}
\value{
list containing the statistic and the p-value of global test
}
\description{
F-Global test.
}
\details{
Internal function. \code{Fg.test.pls} is called by \code{test.partition.pls}.
}
\keyword{internal}
