% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal_F_tests.R
\name{Fg.test}
\alias{Fg.test}
\title{F-global test}
\usage{
Fg.test(Y0, X0, Y1, X1)
}
\arguments{
\item{Y0}{vector of the concatenate indipendent latent variables of H0 hypothesis global 
test}

\item{X0}{matrix of the concatenate predictor latent variables of H0 hypothesis global test}

\item{Y1}{vector of the concatenate indipendent latent variables of H1 hypothesis global 
test}

\item{X1}{matrix of the concatenate predictor latent variables of H1 hypothesis global test}
}
\value{
a list containing the statistic and the p-value obtained by applying F-global test
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
