% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Fg.test.reg}
\alias{Fg.test.reg}
\title{Global test for the regression model}
\usage{
Fg.test.reg(Y0, X0, Y1, X1, method, ...)
}
\arguments{
\item{Y0}{vector of the concatenate indipendent latent variables of null hypothesis global test.}

\item{X0}{matrix of the concatenate predictor latent variables of null hypothesis global test.}

\item{Y1}{vector of the concatenate indipendent latent variables of alternative hypothesis global test.}

\item{X1}{matrix of the concatenate predictor latent variables of alternative hypothesis global test.}

\item{method}{string indicating the method: LM or LAD}

\item{\dots}{Further arguments passed on to \code{\link{Fg.test.reg}}.}
}
\value{
list containing the statistic and the p-value of global test
}
\description{
Global test for the regression model
}
\details{
Internal function. \code{Fg.test.reg} is called by \code{test.partition.reg}.
}
\keyword{internal}

