% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fglobal.tree}
\alias{fglobal.tree}
\title{Listing the the f-global test results for regression model for each obtained partition}
\usage{
fglobal.tree(tree, ...)
}
\arguments{
\item{tree}{class containing the tree element.}

\item{\dots}{Further arguments passed on to \code{\link{fglobal.tree}}.}
}
\value{
the f-global test results for each obtained partition
}
\description{
Listing the the f-global test results for regression model for each obtained partition
}
\details{
Internal function. \code{fglobal.tree.reg} is called by \code{reg.pathmox}.
}
\keyword{internal}

