% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{fibtelereg}
\alias{fibtelereg}
\title{Fibtelereg}
\format{A data frame with 147 observations on the following 18 variables. The first ten variables
are segmentation variables. The rest of the variables refer to five variables 1)
\code{Image} = Image, 2) \code{Exp.spec} = Specific Expectation, 3) \code{Exp.gen} = Generic Expectation,
4)\code{Qual.spec} = Specific Quality, 5) \code{Qual.gen} = Generic Quality, 6) \code{Value} = Value, 7)
\code{Satis} = Satisfaction.
 Variables description
\itemize{
\item{\code{Image}}: Generic students perception of ICT schools: (internationally recognized,
     ranges of courses, leader in research).
\item{\code{Exp.spec}}: Specific Expectation on specific skills (technic or applied skills).
\item{\code{Exp.gen}}: Generic Expectation on generic skills (abilities in problem solving,
 communication skills).
\item{\code{Qual.spec}}: Perception about the achieved quality on the specific skills in the school.
\item{\code{Qual.gen}}: Perception about achieved quality on the  generic skills in
the school (abilities in solving problem, communication skills).
\item{\code{Value}}: The advantage or profit that the alumni may draw from the school
degree (well paid job, motivated job, prospectives in improvement and promotion).
\item{\code{Satis}}: Degree of alumni satisfaction about the formation in school respect to
their actual work conditions.
}

Segmentation Variables description
\itemize{
\item{\code{Career}}{a factor with levels \code{EI} \code{ETS} \code{TEL}}
\item{\code{Gender}}{a factor with levels \code{female} \code{male}}
\item{\code{Age}}{a factor with levels \code{25-26years} \code{27-28years} \code{29-30years} \code{31years+}}
\item{\code{Studying}}{a factor with levels \code{no.stud} \code{yes.stud}}
\item{\code{Contract}} {a factor with levels \code{fix.cont} \code{other.cont} \code{temp.cont}}
\item{\code{Salary}}{a factor with levels \code{18k} \code{>45k} \code{25k} \code{35k} \code{45k}}
\item{\code{Firmtype}}{a factor with levels \code{priva} \code{publi}}
\item{\code{Accgrade}}{a factor with levels \code{7-8accnote} \code{ accnote<7} \code{accnote>8}}
\item{\code{Grade}}{a factor with levels \code{<6.5note} \code{>7.5note} \code{6.5-7note} \code{7-7.5note}}
\item{\code{Startwork}}{a factor with levels \code{after.grad} \code{befor.grad}}
}}
\source{
Laboratory of Information Analysis and Modeling (LIAM).
   Facultat de Informatica de Barcelona, Universitat Politecnica de Catalunya.
}
\usage{
fibtelereg
}
\description{
Fibtelereg dataset
}
\references{
Lamberti, G. (2014) \emph{Modeling with Heterogeneity.} PhD Dissertation.
}
\keyword{datasets}

