% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_turner.R
\name{from_to}
\alias{from_to}
\title{Starting and ending positions}
\usage{
from_to(x, ...)
}
\arguments{
\item{x}{a numeric vector or a list of vectors}

\item{...}{further arguments are ignored}
}
\value{
A list with two vectors: '$from' and '$to'.
'$from' contains the indices with starting positions.
'$to' contains the indices with ending positions.
}
\description{
Starting and ending positions
}
\details{
Internal function.
}
\keyword{internal}
