% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_turner.R
\name{funlist}
\alias{funlist}
\title{Apply a function to all elements in a list}
\usage{
funlist(alist, f, ...)
}
\arguments{
\item{alist}{a list}

\item{f}{a function to be applied}

\item{...}{further arguments passed on to \code{f}}
}
\value{
value
}
\description{
Apply a function to all elements in a list
}
\details{
Internal function.
}
\keyword{internal}
