% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{get_boot_stats}
\alias{get_boot_stats}
\title{Get data frame with bootstrap statistics}
\usage{
get_boot_stats(MATRIX, original)
}
\arguments{
\item{MATRIX}{Matrix with bootstrapped results}

\item{original}{vector with original values}
}
\description{
Get data frame with bootstrap statistics
}
\details{
Internal function. \code{get_boot_stats} is called by \code{get_boots}.
}
\keyword{internal}
