% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{get_boots}
\alias{get_boots}
\title{Performs bootstrap validation in \code{plspm}}
\usage{
get_boots(DM, path_matrix, blocks, specs, br)
}
\arguments{
\item{DM}{Data Matrix}

\item{path_matrix}{Inner Design Matrix}

\item{blocks}{list of vectors with column indices}

\item{specs}{list with algorithm specifications}

\item{br}{number of bootstrap resamples}
}
\description{
Performs bootstrap validation in \code{plspm}
}
\details{
Internal function. \code{get_boots} is called by \code{plspm}.
}
\keyword{internal}
