% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{get_dummies}
\alias{get_dummies}
\title{Dummy matrices for categorical manifest variables}
\usage{
get_dummies(MV, specs)
}
\arguments{
\item{MV}{matrix or data frame from with manifest variables}

\item{specs}{list with algorithm specifications}
}
\value{
dummies list with dummy matrices for categorical manifest variables
}
\description{
Dummy matrices for categorical manifest variables
}
\details{
Internal function. \code{get_dummies} is called by \code{plspm}.
}
\note{
If there are no categorical variables specified in specs,
\code{get_dummies} returns \code{NULL}
}
\keyword{internal}
