% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{get_effects}
\alias{get_effects}
\title{Path coefficient effects for \code{plspm}}
\usage{
get_effects(Path)
}
\arguments{
\item{Path}{matrix of path coefficients}
}
\value{
data frame with relationships, direct, indirect, and total effects
}
\description{
Path coefficient effects for \code{plspm}
}
\details{
Internal function. \code{get_effects} is called by \code{plspm}.
}
\keyword{internal}
