% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{get_gof}
\alias{get_gof}
\title{Goodness-of-fit for \code{plspm}}
\usage{
get_gof(comu, R2, blocks, path_matrix)
}
\arguments{
\item{comu}{list of communalities}

\item{R2}{vector of R-squared coefficients}

\item{blocks}{list of variables in each block}

\item{path_matrix}{Inner Design Matrix}
}
\description{
Goodness-of-fit for \code{plspm}
}
\details{
Internal function. \code{get_gof} is called by \code{plspm}
}
\keyword{internal}
