% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{get_numerics}
\alias{get_numerics}
\title{Transform factors in MV into numeric}
\usage{
get_numerics(MV)
}
\arguments{
\item{MV}{data frame with manifest variables as factors}
}
\value{
list with transformed MV and categories of factors
}
\description{
Transform factors in MV into numeric
}
\details{
Internal function. \code{get_numerics} is called by 
\code{get_treated_data}.
}
\note{
This function is used when there are categorical variables 
defined as 'factors'. The idea is to convert those factors into numeric
vectors while keeping the levels (ie categories) in a separate list
}
\keyword{internal}
