% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{get_ord_scale}
\alias{get_ord_scale}
\title{Calculate inner weighting path scheme}
\usage{
get_ord_scale(y, x, Xdummy)
}
\arguments{
\item{y}{vector of values}

\item{x}{vector of the natural number series 1, 2, ..., p obtained 
by means of the function 'get_rank'}

\item{Xdummy}{dummy matrix corresponding to x (obtained with get_dummy)}

\item{path_matrix}{matrix with paths}

\item{LV}{matrix of latent variables}
}
\value{
scaled matrix
}
\description{
Calculate inner weighting path scheme

Non-Metric Ordinal Scale
}
\details{
Internal function. \code{get_path_scheme} is called by \code{get_weights}

Internal function. \code{get_ord_scale} is called by \code{plspm}.
}
\note{
This function replaces the elements of x by the the means of y conditioned 
to the levels of x
}
\keyword{internal}
