% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{get_path_scheme}
\alias{get_path_scheme}
\title{Calculate inner weighting path scheme}
\usage{
get_path_scheme(path_matrix, LV)
}
\arguments{
\item{path_matrix}{matrix with paths}

\item{LV}{matrix of latent variables}
}
\description{
Calculate inner weighting path scheme
}
\details{
Internal function. \code{get_path_scheme} is called by \code{get_weights}
}
\keyword{internal}
