% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{get_paths}
\alias{get_paths}
\title{Calculate path coefficients for \code{plspm}}
\usage{
get_paths(path_matrix, Y_lvs, full = TRUE)
}
\arguments{
\item{path_matrix}{path matrix}

\item{Y_lvs}{Matrix of latent variables}

\item{full}{logical to indicate all results from 'summary(lm())'}
}
\value{
list with inner results, path coefs matrix, R2, and residuals
}
\description{
Calculate path coefficients for \code{plspm}
}
\details{
Internal function. \code{get_paths} is called by \code{plspm}.
}
\keyword{internal}
