% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{get_PLSR}
\alias{get_PLSR}
\title{Internal PLS regression (full data)}
\usage{
get_PLSR(Y, X, ncomp)
}
\arguments{
\item{Y}{an (already centered) vector of order n}

\item{X}{an (already centered) n x p matrix}

\item{ncomp}{number of components}
}
\description{
Internal PLS regression (full data)
}
\details{
Internal function. \code{get_PLSR}
}
\keyword{internal}
