% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{get_PLSR_NA}
\alias{get_PLSR_NA}
\title{Internal PLS regression with missing values}
\usage{
get_PLSR_NA(Y, X, ncomp)
}
\arguments{
\item{Y}{an (already centered) vector of order n; \code{NA} not allowed}

\item{X}{an (already centered) n x p matrix; \code{NA} allowed but
with at least one observation for every row and column}

\item{ncomp}{number of components}
}
\description{
Internal PLS regression with missing values
}
\details{
Internal function. \code{get_PLSR_NA}
}
\keyword{internal}
