% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{get_scores}
\alias{get_scores}
\title{Calculate Latent Variable Scores}
\usage{
get_scores(X, W)
}
\arguments{
\item{X}{matrix of scaled/quantified manifest variables}

\item{W}{matrix with outer weights (ODM: Outer Design Matrix)}
}
\value{
Scores matrix of latent variables scores
}
\description{
Calculate Latent Variable Scores
}
\details{
Internal function. \code{get_scores} is called by \code{plspm}.
}
\keyword{internal}
