% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{get_treated_data}
\alias{get_treated_data}
\title{Apply corresponding treatment to MV}
\usage{
get_treated_data(MV, specs)
}
\arguments{
\item{MV}{matrix or data frame from with manifest variables}

\item{specs}{list with algorithm specifications}
}
\value{
matrix or data frame of treated MVs (scaling)
}
\description{
Apply corresponding treatment to MV
}
\details{
Internal function. \code{get_treated_data} is called by \code{plspm}.
}
\note{
In non-metric case, all mvs are starndardized, and those ordinal or nominal
are rankified
}
\keyword{internal}
