% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{get_unidim}
\alias{get_unidim}
\title{Unidimensionality of reflective blocks}
\usage{
get_unidim(DM, blocks, modes)
}
\arguments{
\item{DM}{Data Matrix}

\item{blocks}{vector with numbers of variables per block}

\item{modes}{vector of modes}
}
\value{
A data frame with the following columns:

\item{Type.measure}{Measurement mode}

\item{MVs}{number of manifest variables in each block}

\item{C.alpha}{Cronbach's alpha}

\item{DG.rho}{Dillon-Goldstein rho}

\item{eig.1st}{First eigenvalue}

\item{eig.2nd}{Second eigenvalue}
}
\description{
Unidimensionality of reflective blocks
}
\details{
Internal function. \code{get_unidim} is called by \code{plspm}.
}
\keyword{internal}
