% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{has_dimension}
\alias{has_dimension}
\title{Has dimension?}
\usage{
has_dimension(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Has dimension?
}
\details{
Internal function.
}
\keyword{internal}
