% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{has_missing}
\alias{has_missing}
\title{Has missing values, NA, NaN, Inf}
\usage{
has_missing(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Has missing values, NA, NaN, Inf
}
\details{
Internal function.
}
\keyword{internal}
