% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{has_rownames}
\alias{has_rownames}
\title{Has or lacks row/column names?}
\usage{
has_rownames(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Has or lacks row/column names?
}
\details{
Internal function.
}
\keyword{internal}
