% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_turner.R
\name{indexify}
\alias{indexify}
\title{Create indices for elements in a vector or list}
\usage{
indexify(x, out)
}
\arguments{
\item{x}{a numeric vector or list of vectors}

\item{out}{string indicating the output format 
(\code{"vector"} or \code{"list"})}
}
\value{
A vector (or list) of indexed numbers
}
\description{
Create indices for elements in a vector or list
}
\details{
Internal function.
}
\keyword{internal}
