% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_class_and_method.R
\docType{class}
\name{info-class}
\alias{info-class}
\title{info class}
\description{
info is an S4 class that contains info about the tree partition
and the the the Fischer statitistcs: F-global and F-coefficientes
}
\keyword{internal}
