% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal_tree.R
\name{info.mox}
\alias{info.mox}
\title{General information about the pathmox algorithm}
\usage{
info.mox(signif, size, deep, y)
}
\arguments{
\item{signif}{stop condition 1: significance of the p-value}

\item{size}{stop condition 2: minimum number of individuals in a node}

\item{deep}{stop condition 3: maximum tree depth level}

\item{y:}{set of segmentation variables}
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
