% Generated by roxygen2 (4.0.2): do not edit by hand
\name{info.mox.reg}
\alias{info.mox.reg}
\title{Giving general information about the pathmox-lm algorithm}
\usage{
info.mox.reg(signif, size, deep, y, method, ...)
}
\arguments{
\item{signif}{stop condition 1: significance of the p-value.}

\item{size}{stop condition 2: minimum number of individuals  in a node.}

\item{deep}{stop condition 3: maximum tree depth level.}

\item{y}{set of segmentation variables.}

\item{method}{string indicating the method: LM or LAD.}

\item{\dots}{Further arguments passed on to \code{\link{info.mox.reg}}.}
}
\description{
Giving general information about the pathmox-lm algorithm
}
\details{
Internal function. \code{info.mox.reg} is called by \code{reg.pathmox}.
}
\keyword{internal}

