% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class1_info.pls.R
\docType{class}
\name{info.pls-class}
\alias{info.pls-class}
\title{info.pls class}
\description{
info.pls is an S4 class that contains info on the variable and his levels 
that provides the best binary split and the the the Fischers statitistcs: 
F-global, F-#' block, F-coefficientes
}
