% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{info.pls_class}
\alias{info.pls_class}
\title{info.pls class}
\description{
info.pls is a S4 class that contains info on the variable and his levels that
provides the best binary split and the the the Fischers statitistcs: F-global,
F- block, F-coefficientes
}

