% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class4_info.reg.R
\docType{class}
\name{info.reg_class}
\alias{info.reg_class}
\title{info.reg class}
\description{
info.pls is a S4 class that contains info on the variable and his levels that
provides the best binary split and the the the Fischers statitistcs: F-global,
F-coefficientes
}
