% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_class}
\alias{is_class}
\title{Is class}
\usage{
is_class(x, name = NULL)
}
\arguments{
\item{x}{an R object}

\item{name}{string giving the class to be tested}
}
\description{
Is class
}
\details{
Internal function.
}
\keyword{internal}
