% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_decimal}
\alias{is_decimal}
\title{Is decimal}
\usage{
is_decimal(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Is decimal
}
\details{
Internal function.

decimal is any number in the intervals (-1,0) and (0,1)
}
\keyword{internal}
