% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_diagonal}
\alias{is_diagonal}
\title{Is diagonal matrix}
\usage{
is_diagonal(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Is diagonal matrix
}
\details{
Internal function.
}
\keyword{internal}
