% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_lower_triangular}
\alias{is_lower_triangular}
\title{Is triangular matrix}
\usage{
is_lower_triangular(x, diag = FALSE)
}
\arguments{
\item{x}{a matrix}

\item{diag}{should the diagonal be included? (\code{FALSE} by default)}
}
\description{
Is triangular matrix
}
\details{
Internal function.
}
\keyword{internal}
