% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{is_missing}
\alias{is_missing}
\title{Presence of missing values}
\usage{
is_missing(x)
}
\arguments{
\item{x}{a list}
}
\value{
vector indicating whether each element in x contains NAs
}
\description{
Presence of missing values
}
\details{
Internal function. \code{is_missing} is called by 
\code{get_weights_nonmetric}
}
\keyword{internal}
