% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_multidim}
\alias{is_multidim}
\title{Test if an object is multi-dimensional}
\usage{
is_multidim(x)
}
\arguments{
\item{x}{an R object}
}
\value{
whether x is multi-dimensional
}
\description{
Test if an object is multi-dimensional
}
\details{
Internal function.
}
\keyword{internal}
