% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_multiple}
\alias{is_multiple}
\title{Is multiple}
\usage{
is_multiple(x, of)
}
\arguments{
\item{x}{a numeric object}

\item{of}{a given number}
}
\description{
Is multiple
}
\details{
Internal function.
}
\keyword{internal}
