% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_negative_decimal}
\alias{is_negative_decimal}
\title{Is negative decimal}
\usage{
is_negative_decimal(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Is negative decimal
}
\details{
Internal function.
}
\keyword{internal}
