% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_negative_integer}
\alias{is_negative_integer}
\title{Is negative integer}
\usage{
is_negative_integer(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Test if is a positive integer
}
\keyword{internal}
