% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_one_dim}
\alias{is_one_dim}
\title{Test if an object has one-dimension}
\usage{
is_one_dim(x)
}
\arguments{
\item{x}{an R object}
}
\value{
whether x is one-dimensional
}
\description{
Test if an object has one-dimension
}
\details{
Internal function.
}
\keyword{internal}
