% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_rectangular_matrix}
\alias{is_rectangular_matrix}
\title{Is rectangular matrix}
\usage{
is_rectangular_matrix(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Is rectangular matrix
}
\details{
Internal function.
}
\keyword{internal}
