% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_single_decimal}
\alias{is_single_decimal}
\title{Is single decimal}
\usage{
is_single_decimal(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Is single decimal
}
\details{
Internal function.
}
\keyword{internal}
