% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_single_even}
\alias{is_single_even}
\title{Is single even}
\usage{
is_single_even(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Is single even
}
\details{
Internal function.
}
\keyword{internal}
