% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_single_logical}
\alias{is_single_logical}
\title{Is single logical}
\usage{
is_single_logical(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Is single logical
}
\details{
Internal function.
}
\keyword{internal}
