% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_single_negative}
\alias{is_single_negative}
\title{Is single negative number}
\usage{
is_single_negative(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Is single negative number
}
\details{
Internal function.
}
\keyword{internal}
