% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_single_negative_integer}
\alias{is_single_negative_integer}
\title{Is single negative integer}
\usage{
is_single_negative_integer(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Is single negative integer
}
\details{
Internal function.
}
\keyword{internal}
