% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_single_number}
\alias{is_single_number}
\title{Is single number}
\usage{
is_single_number(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Is single number
}
\details{
Internal function.
}
\keyword{internal}
