% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_single_positive}
\alias{is_single_positive}
\title{Is single positive number}
\usage{
is_single_positive(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Is single positive number
}
\details{
Internal function.
}
\keyword{internal}
