% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_single_positive_decimal}
\alias{is_single_positive_decimal}
\title{Is single positive decimal}
\usage{
is_single_positive_decimal(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Is single positive decimal
}
\details{
Internal function.
}
\keyword{internal}
