% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_single_positive_integer}
\alias{is_single_positive_integer}
\title{Is single positive integer}
\usage{
is_single_positive_integer(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Is single positive integer
}
\details{
Internal function.
}
\keyword{internal}
