% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_single_string}
\alias{is_single_string}
\title{Is single string}
\usage{
is_single_string(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Is single string
}
\details{
Internal function.
}
\keyword{internal}
