% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_single_true}
\alias{is_single_true}
\title{Is single true}
\usage{
is_single_true(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Is single true
}
\details{
Internal function.
}
\keyword{internal}
