% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{is_square_numeric_matrix}
\alias{is_square_numeric_matrix}
\title{Is square numeric matrix}
\usage{
is_square_numeric_matrix(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Is square numeric matrix
}
\details{
Internal function.
}
\keyword{internal}
