% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_turner.R
\name{lengths}
\alias{lengths}
\title{Length of each element within a list}
\usage{
lengths(alist, out = "vector")
}
\arguments{
\item{alist}{a list}

\item{out}{string indicating the format of the output (\code{"vector"} or 
\code{"list"})}
}
\value{
A vector (or list) with the lengths of the elements in \code{alist}
}
\description{
Length of each element within a list
}
\details{
Internal function.
}
\keyword{internal}
