% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{list_of_vectors}
\alias{list_of_vectors}
\title{List of vectors}
\usage{
list_of_vectors(x)
}
\arguments{
\item{x}{an R object}
}
\description{
List of vectors
}
\details{
Internal function.
}
\keyword{internal}
