% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_turner.R
\name{list_ones}
\alias{list_ones}
\title{List with vectors of ones}
\usage{
list_ones(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
A list of vectors with ones
}
\description{
List with vectors of ones
}
\details{
Internal function.
}
\keyword{internal}
