% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_turner.R
\name{list_to_dummy}
\alias{list_to_dummy}
\title{Create a dummy matrix from the elements in a list}
\usage{
list_to_dummy(alist)
}
\arguments{
\item{alist}{a list of vectors}
}
\value{
A matrix of dummy variables
}
\description{
Create a dummy matrix from the elements in a list
}
\details{
Internal function.
}
\keyword{internal}
