% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_turner.R
\name{list_to_matrix}
\alias{list_to_matrix}
\title{Design-type matrix from the elements in a list}
\usage{
list_to_matrix(alist)
}
\arguments{
\item{alist}{a list of numeric vectors}
}
\value{
A design-type matrix
}
\description{
Design-type matrix from the elements in a list
}
\details{
Internal function.
}
\keyword{internal}
