% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{list_with_vectors}
\alias{list_with_vectors}
\title{List with vectors}
\usage{
list_with_vectors(x)
}
\arguments{
\item{x}{an R object}
}
\description{
List with vectors
}
\details{
Internal function.
}
\keyword{internal}
