% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_turner.R
\name{listify}
\alias{listify}
\title{Create a list from a vector of integers}
\usage{
listify(indices)
}
\arguments{
\item{indices}{a vector of integers indicating the length of each vector
in the produced list}
}
\value{
A list of index vectors
}
\description{
Create a list from a vector of integers
}
\details{
Internal function.
}
\keyword{internal}
