% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_turner.R
\name{listsize}
\alias{listsize}
\title{Size: total number of elements in a list}
\usage{
listsize(alist)
}
\arguments{
\item{alist}{a list}
}
\value{
number of elements in \code{alist}.
}
\description{
Size: total number of elements in a list
}
\details{
Internal function.
}
\keyword{internal}
