% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_turner.R
\name{matrix_to_blocks}
\alias{matrix_to_blocks}
\title{Split a matrix into blocks}
\usage{
matrix_to_blocks(Matrix, blocks, byrow = TRUE)
}
\arguments{
\item{Matrix}{a matrix to split}

\item{blocks}{either a list or a vector indicating the blocks. 
If \code{blocks} is a list of vectors, then the length of each vector 
defines the size of the blocks. If \code{blocks} is a vector, then each 
element represents the size of the blocks.}

\item{byrow}{logical. If \code{TRUE} (the default) the matrix is split 
by rows, otherwise the matrix is split by columns}
}
\value{
A list of matrices
}
\description{
Split a matrix into blocks
}
\details{
Internal function.
}
\keyword{internal}
