% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_turner.R
\name{maxlist}
\alias{maxlist}
\title{Maximum of all elements in a list}
\usage{
maxlist(alist, na.rm = FALSE)
}
\arguments{
\item{alist}{a list}

\item{na.rm}{logical indicating whether missing values should be removed}
}
\value{
the maximum
}
\description{
Maximum of all elements in a list
}
\details{
Internal function.
}
\keyword{internal}
