% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25_fun_mox.tree.pls.R
\name{mox.tree.pls}
\alias{mox.tree.pls}
\title{Resuming the tree attributes: it gives general information about the tree}
\usage{
mox.tree.pls(tree, ...)
}
\arguments{
\item{tree}{class containing the tree element.}

\item{\dots}{Further arguments passed on to \code{\link{mox.tree.pls}}.}
}
\value{
a data with general information about the tree and its nodes
}
\description{
Resuming the tree attributes: it gives general information about the tree
}
\details{
Internal function. \code{mox.tree} is called by \code{pls.pathmox}.
}
\keyword{internal}
