% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_fun_mox.tree.R
\name{mox.tree}
\alias{mox.tree}
\title{Resuming the tree attributes (regression model) : it gives general information about the tree}
\usage{
mox.tree(tree, ...)
}
\arguments{
\item{tree}{class containing the tree element.}

\item{\dots}{Further arguments passed on to \code{\link{mox.tree}}.}
}
\value{
a data with general information about the tree and its nodes
}
\description{
Resuming the tree attributes (regression model) : it gives general information about the tree
}
\details{
Internal function. \code{mox.tree.reg} is called by \code{reg.pathmox}.
}
\keyword{internal}
