% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{normalize}
\alias{normalize}
\title{Normalize a vector}
\usage{
normalize(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
normalized vector
}
\description{
Normalize a vector
}
\details{
Internal function. \code{normalize} is called internally
}
\keyword{internal}
