% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_fun_partition.R
\name{partition}
\alias{partition}
\title{Defining binary partitions given a segmentation variable (factor).}
\usage{
partition(x, ...)
}
\arguments{
\item{x}{single factor or a data.frame of segmenation variables.}

\item{\dots}{Further arguments passed on to \code{\link{partition}}.}
}
\value{
list of matrices containing the all possibiles binary partions given a segmenation variable.
}
\description{
Defining binary partitions given a segmentation variable (factor).
}
\details{
Internal function. \code{partition} is called by \code{all.part.pls}.
}
\keyword{internal}
