% Generated by roxygen2 (4.0.2): do not edit by hand
\name{partopt.reg}
\alias{partopt.reg}
\title{Defining the optimum partition given a set of segmentation variables (regression model)}
\usage{
partopt.reg(x, y, method, ...)
}
\arguments{
\item{x}{matrix or data.frame with the data.}

\item{y}{matrix or data.frame of the segmentation variables.}

\item{method}{string indicating the method: LM or LAD}

\item{\dots}{Further arguments passed on to \code{\link{all.part.reg}}.}
}
\value{
list containing information of the optimum partition given a set of segmentation variables
}
\description{
Defining the optimum partition given a set of segmentation variables (regression model)
}
\details{
Internal function. \code{partopt.reg} is called by \code{reg.pathmox}.
}
\keyword{internal}

