% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal_tree.R
\name{percent.node}
\alias{percent.node}
\title{Calculating size (numeber of individual of a node)}
\usage{
percent.node(x, size)
}
\arguments{
\item{x}{matrix or dataframe with data.}

\item{size}{value indicating the minimun threshold of number of observations for a node}
}
\value{
the number of observations in a node
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
