% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{pls_paths}
\alias{pls_paths}
\title{Calculate path coefficients for \code{pls}}
\usage{
pls_paths(inner, Y_lvs)
}
\arguments{
\item{inner}{path matrix}

\item{Y_lvs}{matrix of latent variables}
}
\value{
a matrix with path coefs and R2
}
\description{
############################################################################################
}
\details{
Internal function.
}
\keyword{internal}
