% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{pls_weights}
\alias{pls_weights}
\title{Outer Weights}
\usage{
pls_weights(X, inner, blocks, modes, scheme, tol, iter)
}
\arguments{
\item{X}{scaled data}

\item{inner}{matrix with path connections}

\item{blocks}{list with variables in each block}

\item{modes}{character vector indicating the type of measurement for each
block. Possible values are: \code{"A", "B"} 
The length of \code{modes} must be equal to the length of \code{blocks}}

\item{scheme}{string indicating the type of inner weighting
scheme. Possible values are \code{"centroid"}, \code{"factorial"}, or
\code{"path"}}

\item{tol}{decimal value indicating the tolerance criterion for the
iterations (\code{tol=0.000001}). Can be specified between 0 and 0.001}

\item{iter}{integer indicating the maximum number of iterations
(\code{maxiter=100} by default). The minimum value of \code{maxiter} is 100}
}
\value{
a vector with outer weights
}
\description{
############################################################################################
}
\details{
Internal function
}
\note{
Calculate outer weights (under Lohmoller's algorithm)
}
\keyword{internal}
