% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_class_and_method.R
\name{plstree}
\alias{plstree}
\title{create method plstree}
\usage{
plstree(x, ...)
}
\arguments{
\item{x}{the element representing the method.}

\item{\dots}{Further arguments passed on to \code{\link{plstree}}.}
}
\value{
internal method
}
\description{
create method plstree
}
\details{
Internal function.
}
\keyword{internal}
