% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal_tree.R
\name{printTree}
\alias{printTree}
\title{printing the tree structure}
\usage{
printTree(moxtree)
}
\arguments{
\item{moxtree}{moxtree object}
}
\value{
the tree structure
}
\description{
############################################################################################
}
\details{
Internal function.
}
\keyword{internal}
