% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_turner.R
\name{prodlist}
\alias{prodlist}
\title{Product of all elements in a list}
\usage{
prodlist(alist, na.rm = FALSE)
}
\arguments{
\item{alist}{a list}

\item{na.rm}{logical indicating whether missing values should be removed}
}
\value{
the product
}
\description{
Product of all elements in a list
}
\details{
Internal function.
}
\keyword{internal}
