% Generated by roxygen2 (4.0.2): do not edit by hand
\name{root.tree}
\alias{root.tree}
\title{Calculating the element of the root node}
\usage{
root.tree(tree, ...)
}
\arguments{
\item{tree}{class containing the tree element}

\item{\dots}{Further arguments passed on to \code{\link{root.tree}}.}
}
\value{
the elements of the root node
}
\description{
Calculating the element of the root node
}
\details{
Internal function. \code{root.tree} is called by \code{pls.pathmox}.
}
\keyword{internal}

