% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{same_class}
\alias{same_class}
\title{Same Class}
\usage{
same_class(x, y)
}
\arguments{
\item{x}{an R object}

\item{y}{an R object}
}
\description{
Same Class
}
\details{
Internal function.
}
\keyword{internal}
