% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{same_dim}
\alias{same_dim}
\title{Same Dimension}
\usage{
same_dim(x, y)
}
\arguments{
\item{x}{a matrix}

\item{y}{a matrix}
}
\description{
Same Dimension
}
\details{
Internal function.
}
\keyword{internal}
