% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{same_length}
\alias{same_length}
\title{Same Length}
\usage{
same_length(x, y)
}
\arguments{
\item{x}{a matrix}

\item{y}{a matrix}
}
\description{
Same Length
}
\details{
Internal function.
}
\keyword{internal}
