% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{same_mode}
\alias{same_mode}
\title{Same Mode}
\usage{
same_mode(x, y)
}
\arguments{
\item{x}{an R object}

\item{y}{an R object}
}
\description{
Same Mode
}
\details{
Internal function.
}
\keyword{internal}
