% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{same_nrow}
\alias{same_nrow}
\title{Same Number of Rows / Columns}
\usage{
same_nrow(x, y)
}
\arguments{
\item{x}{a matrix}

\item{y}{a matrix}
}
\description{
Same Number of Rows / Columns
}
\details{
Internal function.
}
\keyword{internal}
