% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tester.R
\name{same_type}
\alias{same_type}
\title{Same Type}
\usage{
same_type(x, y)
}
\arguments{
\item{x}{an R object}

\item{y}{an R object}
}
\description{
Same Type
}
\details{
Internal function.
}
\keyword{internal}
