% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal_tree.R
\name{showDeepth}
\alias{showDeepth}
\title{Calculating Deepth stop criterion}
\usage{
showDeepth(node)
}
\arguments{
\item{node}{id that identifies a specicif node}
}
\value{
deepth of the tree
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
