% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathmox_internal.R
\name{splitopt}
\alias{splitopt}
\title{Best partition for a specific segmentation variable}
\usage{
splitopt(x, inner, outer, mode, scheme, splits, fact, size_candidate)
}
\arguments{
\item{x}{matrix or data frame containing the manifest variables}

\item{inner}{a square (lower triangular) boolean matrix representing 
the inner model (i.e. the path relationships between latent variables)}

\item{outer}{list of vectors with column indices or column names
from \code{Data} indicating the sets of manifest variables forming 
each block (i.e. which manifest variables correspond to each block)}

\item{mode}{character vector indicating the type of measurement for each
block. Possible values are: \code{"A", "B"} 
The length of \code{modes} must be equal to the length of \code{blocks}}

\item{scheme}{string indicating the type of inner weighting
scheme. Possible values are \code{"centroid"}, \code{"factorial"}, or
\code{"path"}}

\item{splits}{vector indicating the binary partition}

\item{fact}{vector indicating the categorical variable}

\item{size_candidate}{number indicating the minimum threshold for a node}
}
\value{
a list containing information about the bets partition for a specific segmentation 
variable
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
