% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_turner.R
\name{sumlist}
\alias{sumlist}
\title{Sum of all elements in a list}
\usage{
sumlist(alist, na.rm = FALSE)
}
\arguments{
\item{alist}{a list}

\item{na.rm}{logical indicating whether missing values should be removed}
}
\value{
the sum
}
\description{
Sum of all elements in a list
}
\details{
Internal function.
}
\keyword{internal}
