% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal_tree.R
\name{terminal.tree}
\alias{terminal.tree}
\title{Observations belonging to the terminal nodes}
\usage{
terminal.tree(moxtree)
}
\arguments{
\item{moxtree}{class containing the moxtree element.}
}
\value{
the observations belonging to the terminal nodes
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
