% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal_F_tests.R
\name{test.partition}
\alias{test.partition}
\title{Cheking F-tests results}
\usage{
test.partition(x, inner, .model, .scheme, .consistent, modtwo, alpha)
}
\arguments{
\item{x}{matrix or dataframe containing the data.}

\item{inner}{a square (lower triangular) boolean matrix representing the inner model.}

\item{.model}{A description of the user-specified model.}

\item{.scheme}{string indicating the type of inner weighting
scheme. Possible values are \code{"centroid"}, \code{"factorial"}, or \code{"path"}.}

\item{.consistent}{Logical. Should composite/proxy correlations be disattenuated 
to yield consistent loadings and path estimates if at least one of the construct 
is modeled as a common factor. Defaults to TRUE.}

\item{modtwo}{vector indicating the binary partition.}

\item{alpha}{value indicating the threshold a stop condition.}
}
\value{
list containing matrices needed for the comparison test.
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
