% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{test_factors}
\alias{test_factors}
\title{Test presence of factors}
\usage{
test_factors(DF)
}
\arguments{
\item{DF}{data frame that might contain factors}
}
\value{
whether DF contains factors or not
}
\description{
Test presence of factors
}
\details{
Internal function. \code{test_factors} is called by 
\code{get_treated_data}.
}
\note{
Test presence of factors in a data.frame
}
\keyword{internal}
