% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{test_manifest_scaling}
\alias{test_manifest_scaling}
\title{Test scaling of selected manifest variables}
\usage{
test_manifest_scaling(MV, scaling)
}
\arguments{
\item{MV}{matrix or data frame with selected manifest variables}

\item{scaling}{list with specified scaling for each manifest variable}
}
\value{
gives an error when factors in MV have bad scaling
}
\description{
Test scaling of selected manifest variables
}
\details{
Internal function. \code{test_manifest_scaling} is called by \code{plspm}.
Basically checks that unordered factors have nominal scaling.
}
\keyword{internal}
