% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pls.R
\name{test_null_weights}
\alias{test_null_weights}
\title{Test outer weights convergence within specified maxiter}
\usage{
test_null_weights(wgs, specs)
}
\arguments{
\item{wgs}{output from 'get_weights()'}

\item{specs}{list of specifications}
}
\value{
gives an error when wgs is NULL (i.e. non-convergence)
}
\description{
Test outer weights convergence within specified maxiter
}
\details{
Internal function. \code{test_weights_output} is called by \code{plspm}.
Basically checks that weights are not NULL.
}
\keyword{internal}
