% Generated by roxygen2 (4.0.2): do not edit by hand
\name{treemodelreg}
\alias{treemodelreg}
\title{create method treemodelreg}
\usage{
treemodelreg(x, ...)
}
\arguments{
\item{x}{the element representing the method.}

\item{\dots}{Further arguments passed on method.}
}
\value{
internal method
}
\description{
create method treemodelreg
}
\details{
Internal function.
}
\keyword{internal}

