% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathmox_internal.R
\name{var_imp_mox}
\alias{var_imp_mox}
\title{Ranking of variables importance}
\usage{
var_imp_mox(x, y)
}
\arguments{
\item{x}{matrix or dataframe containing the data}

\item{y}{vector or dataframe containing the categorical variables}
}
\value{
a dataframe containg the ranking of the categorical variable
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
