% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_turner.R
\name{vector_to_dummy}
\alias{vector_to_dummy}
\title{Create a dummy matrix from the elements in a vector}
\usage{
vector_to_dummy(avector)
}
\arguments{
\item{avector}{a numeric vector}
}
\value{
A matrix of dummy variables
}
\description{
Create a dummy matrix from the elements in a vector
}
\details{
Internal function.
}
\keyword{internal}
