% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xtree.pls}
\alias{xtree.pls}
\title{create method xtree.pls}
\usage{
xtree.pls(x, ...)
}
\arguments{
\item{x}{the element representing the method.}

\item{\dots}{Further arguments passed on to \code{\link{xtree.pls}}.}
}
\value{
internal method
}
\description{
create method xtree.pls
}
\details{
Internal function.
}
\keyword{internal}

