% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathmox_internal.R
\name{all_part}
\alias{all_part}
\title{Candidates to the bets partition for each of segmentation variables}
\usage{
all_part(x, y, inner, .model, .scheme, .consistent, size_candidate, ...)
}
\arguments{
\item{x}{matrix or dataframe containing the dataset.}

\item{y}{matrix or dataframe or vector of the segmentation variables}

\item{inner}{a square (lower triangular) boolean matrix representing 
the inner model (i.e. the path relationships between latent variables)}

\item{.model}{A description of the user-specified model.}

\item{.scheme}{string indicating the type of inner weighting
scheme. Possible values are \code{"centroid"}, \code{"factorial"}, or
\code{"path"}.}

\item{.consistent}{Logical. Should composite/proxy correlations be disattenuated 
to yield consistent loadings and path estimates if at least one of the construct 
is modeled as a common factor. Defaults to TRUE.}

\item{size_candidate}{number indicating the minimum threshold for a node.}
}
\value{
a list containing information of the candidates to the optimum partition 
for each of segmentation variables
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
