% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathmox_internal.R
\name{bin.levels}
\alias{bin.levels}
\title{Labels of a categorical variable a binary partions}
\usage{
bin.levels(z, spl)
}
\arguments{
\item{z}{matrix containing the posible combination of levels of factor}

\item{variable}{segmentation variables (factor)}
}
\value{
a list of labels
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
