% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathmox_internal.R
\name{blockdiag}
\alias{blockdiag}
\title{Bart matrix}
\usage{
blockdiag(x, ...)
}
\arguments{
\item{x}{list of matrices used to build the Bart matrix}

\item{\dots}{further arguments passed on to \code{\link{blockdiag}}.}
}
\value{
the Bart matrix
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
