% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathmox_internal.R
\name{build.block}
\alias{build.block}
\title{Linear relations between latent variables.}
\usage{
build.block(inner, latent)
}
\arguments{
\item{inner}{a square (lower triangular) boolean matrix representing 
the inner model (i.e. the path relationships between latent variables)}

\item{latent}{matrix of the latent score}
}
\value{
a list of matrices containing for each endogenous latent variable the 
the own predictors
}
\description{
############################################################################################
}
\details{
Internal function.
}
\keyword{internal}
