% Generated by roxygen2 (4.0.2): do not edit by hand
\name{candidates.tree}
\alias{candidates.tree}
\title{Listing the candidates partions for each intermedied nodes}
\usage{
candidates.tree(tree, ...)
}
\arguments{
\item{tree}{class containing the tree element.}

\item{\dots}{Further arguments passed on to \code{\link{candidates.tree}}.}
}
\value{
the candidates partions for intermediated nodes
}
\description{
Listing the candidates partions for each intermedied nodes
}
\details{
Internal function. \code{nodes.tree} is called by \code{pls.pathmox}.
}
\keyword{internal}

