% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathmox_internal.R
\name{check_arg_mox}
\alias{check_arg_mox}
\title{Checks arguments}
\usage{
check_arg_mox(
  .model,
  .data,
  .catvar,
  .scheme,
  .consistent,
  .alpha,
  .deep,
  .size,
  .size_candidate,
  .tree
)
}
\arguments{
\item{.model}{a model in lavaan model syntax}

\item{.data}{a data.frame or a matrix of indicators}

\item{.catvar}{a vector or dataframe containing the categorical variables}

\item{.scheme}{Character string. Approach used to obtain composite weights}

\item{.consistent}{Logical. Should composite/proxy correlations be disattenuated}

\item{.alpha}{minimum threshold of f-test p-value}

\item{.deep}{minimum threshold of deep tree}

\item{.size}{minimum threshold size node}

\item{.size_candidate}{minimum size_candidate node threshold}

\item{.tree}{Logical.Should the tree plot printed}
}
\value{
a list checked arguments
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
