% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathmox_internal.R
\name{element}
\alias{element}
\title{Path coefficient labels}
\usage{
element(x, l)
}
\arguments{
\item{x}{matrix containing information of casual relationship of latent variables}
}
\value{
a vector of path coefficients labels
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
