% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_fun_element.R
\name{element}
\alias{element}
\title{Defining labels of a path coefficient.}
\usage{
element(x, l, ...)
}
\arguments{
\item{x}{matrix containing information of casual relationship of latent variables.}

\item{\dots}{Further arguments passed on to \code{\link{element}}.}
}
\value{
vector path coefficients labels.
}
\description{
Defining labels of a path coefficient.
}
\details{
Internal function. \code{element} is called by \code{F.data.pls}.
}
\keyword{internal}
