% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal_F_tests.R
\name{F.data}
\alias{F.data}
\title{Data preprocessing for F-tests}
\usage{
F.data(x, inner, outer, mode, scheme, modtwo)
}
\arguments{
\item{x}{matrix or dataframe containing the data.}

\item{inner}{a square (lower triangular) boolean matrix representing the inner model}

\item{outer}{list of vectors with column indices indicating which manifest variables 
correspond to each block}

\item{mode}{character vector indicating the type of measurement for each
block. Possible values are: \code{"A", "B"}}

\item{scheme}{string indicating the type of inner weighting
scheme. Possible values are \code{"centroid"}, \code{"factorial"}, or \code{"path"}.}

\item{modtwo}{vector indicating the binary partition}
}
\value{
list containing matrices needed for the tests
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
