% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathmox_internal.R
\name{f.min}
\alias{f.min}
\title{Vector minimum position}
\usage{
f.min(x)
}
\arguments{
\item{x}{vector of values}
}
\value{
a list containing minimun value, the position of the minimun and the 
values different from NA of a vactor
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
