% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_fun_fcoef.tree.reg.R
\name{fcoef.tree.reg}
\alias{fcoef.tree.reg}
\title{Listing the the f-coefficients test results for regression model for each obtained partition}
\usage{
fcoef.tree.reg(tree, ...)
}
\arguments{
\item{tree}{class containing the tree element.}

\item{\dots}{Further arguments passed on to \code{\link{fcoef.tree.reg}}.}
}
\value{
the f-coefficients test results for each obtained partition
}
\description{
Listing the the f-coefficients test results for regression model for each obtained partition
}
\details{
Internal function. \code{fcoef.tree.reg} is called by \code{reg.pathmox}.
}
\keyword{internal}
