% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22_fun_fglobal.tree.pls.R
\name{fglobal.tree.pls}
\alias{fglobal.tree.pls}
\title{Listing the the f-global test results for each obtained partition}
\usage{
fglobal.tree.pls(tree, ...)
}
\arguments{
\item{tree}{class containing the tree element.}

\item{\dots}{Further arguments passed on to \code{\link{fglobal.tree.pls}}.}
}
\value{
the f-global test results for each obtained partition
}
\description{
Listing the the f-global test results for each obtained partition
}
\details{
Internal function. \code{fglobal.tree} is called by \code{pls.pathmox}.
}
\keyword{internal}
