% Generated by roxygen2 (4.0.2): do not edit by hand
\name{info.mox.pls}
\alias{info.mox.pls}
\title{Giving general information about the pathmox algorithm}
\usage{
info.mox.pls(signif, size, deep, y, ...)
}
\arguments{
\item{signif}{stop condition 1: significance of the p-value}

\item{size}{stop condition 2: minimum number of individuals  in a node}

\item{deep}{stop condition 3: maximum tree depth level}

\item{y:}{set of segmentation variables}

\item{\dots}{Further arguments passed on to \code{\link{info.mox.pls}}.}
}
\description{
Giving general information about the pathmox algorithm
}
\details{
Internal function. \code{info.mox.pls} is called by \code{pls.pathmox}.
}
\keyword{internal}

