% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal_tree.R
\name{mox.tree}
\alias{mox.tree}
\title{General information about the tree}
\usage{
mox.tree(moxtree)
}
\arguments{
\item{moxtree}{class containing the tree elements}
}
\value{
a dataframe containing information about the tree and its nodes
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
