% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal_tree.R
\name{nodes.tree}
\alias{nodes.tree}
\title{Observations belonging to the nodes}
\usage{
nodes.tree(moxtree)
}
\arguments{
\item{moxtree}{class containing the moxtree elements}
}
\value{
the observations belonging to the nodes
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
