% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nodes.tree}
\alias{nodes.tree}
\title{Calculating the element of all nodes}
\usage{
nodes.tree(tree, ...)
}
\arguments{
\item{tree}{class containing the tree element}

\item{\dots}{Further arguments passed on to \code{\link{nodes.tree}}.}
}
\value{
the elements of all node
}
\description{
Calculating the element of all nodes
}
\details{
Internal function. \code{nodes.tree} is called by \code{pls.pathmox}.
}
\keyword{internal}

