% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathmox_internal.R
\name{partition}
\alias{partition}
\title{Binary partitions given a segmentation variable (factor).}
\usage{
partition(x)
}
\arguments{
\item{x}{single factor or a data.frame of segmenation variables.}
}
\value{
a list of matrices containing  all possibiles binary partions given a
segmenation variable
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
