% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_fun_percent.node.R
\name{percent.node}
\alias{percent.node}
\title{Calculating size (numeber of individual of a node) stopo criterion}
\usage{
percent.node(x, size, ...)
}
\arguments{
\item{x}{matrix or data.frame with data.}

\item{size}{value of minimun number of cases.}
}
\value{
the minimum number of individuals for a node
}
\description{
Calculating size (numeber of individual of a node) stopo criterion
}
\details{
Internal function. \code{percent.node} is called by \code{pls.pathmox}.
}
\keyword{internal}
