% Generated by roxygen2 (4.0.2): do not edit by hand
\name{printTree}
\alias{printTree}
\title{printing the tree structure}
\usage{
printTree(tree)
}
\arguments{
\item{tree}{that identify the tree object}
}
\value{
the tree structure
}
\description{
printing the tree structure
}
\details{
Internal function. \code{printTree} is called by \code{reg.pathmox}.
}
\keyword{internal}

