% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal_tree.R
\name{root.tree}
\alias{root.tree}
\title{Observations belonging to the root node}
\usage{
root.tree(moxtree)
}
\arguments{
\item{moxtree}{class containing the moxtree elements}
}
\value{
the observations belonging to the root node
}
\description{
############################################################################################
}
\details{
Internal function
}
\keyword{internal}
