% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_fun_showDeepth.R
\name{showDeepth}
\alias{showDeepth}
\title{Calculating Deepth stop criterion}
\usage{
showDeepth(node)
}
\arguments{
\item{node}{id that identify a specicif node}
}
\value{
Deepth of the tree
}
\description{
Calculating Deepth stop criterion
}
\details{
Internal function. \code{showDeepth} is called by \code{pls.pathmox}.
}
\keyword{internal}
