% Generated by roxygen2 (4.0.2): do not edit by hand
\name{showDeepth}
\alias{showDeepth}
\title{Calculating Deepth stop criterion}
\usage{
showDeepth(node)
}
\arguments{
\item{node}{id that identify a specicif node}
}
\value{
Deepth of the tree
}
\description{
Calculating Deepth stop criterion
}
\details{
Internal function. \code{showDeepth} is called by \code{pls.pathmox}.
}
\keyword{internal}

