% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_fun_terminal.tree.R
\name{terminal.tree}
\alias{terminal.tree}
\title{Calculating the element of the terminal nodes}
\usage{
terminal.tree(tree, ...)
}
\arguments{
\item{tree}{class containing the tree element.}

\item{\dots}{Further arguments passed on to \code{\link{terminal.tree}}.}
}
\value{
the elements of the terminal node
}
\description{
Calculating the element of the terminal nodes
}
\details{
Internal function. \code{terminal.tree} is called by \code{pls.pathmox}.
}
\keyword{internal}
