% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/37_fun_test.particion.reg.R
\name{test.particion.reg}
\alias{test.particion.reg}
\title{Calculating the comparison tests for the regression model}
\usage{
test.particion.reg(x, modtwo, signif, method, ...)
}
\arguments{
\item{x}{matrix or data.frame with the data.}

\item{modtwo}{vector indicating the binary partition}

\item{method}{string indicating the method: LM or LAD.}

\item{\dots}{Further arguments passed on to \code{\link{test.particion.reg}}.}

\item{signific}{string indicating a stop condition.}
}
\value{
list containing matrices needed for the comparison test
}
\description{
Calculating the comparison tests for the regression model
}
\details{
Internal function. \code{test.partition.reg} is called by \code{reg.pathmox}.
}
\keyword{internal}
