% Generated by roxygen2 (4.0.2): do not edit by hand
\name{treemodel}
\alias{treemodel}
\title{create method treemodel}
\usage{
treemodel(x, ...)
}
\arguments{
\item{x}{the element representing the method.}

\item{\dots}{Further arguments passed on to \code{\link{treemodel}}.}
}
\value{
internal method
}
\description{
create method treemodel
}
\details{
Internal function.
}
\keyword{internal}

