% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method4_treemodelreg.R
\name{treemodelreg}
\alias{treemodelreg}
\title{create method treemodelreg}
\usage{
treemodelreg(x, ...)
}
\arguments{
\item{x}{the element representing the method.}

\item{\dots}{Further arguments passed on method.}
}
\value{
internal method
}
\description{
create method treemodelreg
}
\details{
Internal function.
}
\keyword{internal}
