% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method3_xtree.reg.R
\name{xtree.reg}
\alias{xtree.reg}
\title{create method xtree.reg}
\usage{
xtree.reg(x, ...)
}
\arguments{
\item{x}{the element representing the method.}

\item{\dots}{Further arguments passed on to \code{\link{xtree.reg}}.}
}
\value{
internal method
}
\description{
create method xtree.reg
}
\details{
Internal function.
}
\keyword{internal}
